+++ webapp/tests/unit/utils/plugin-manager.test.js
@@ -6,8 +6,8 @@
  *                   prevents one PM2 instance's stale in-memory registry from clobbering a peer
  *                   instance's more recent change
  * @file            webapp/tests/unit/utils/plugin-manager.test.js
- * @version         1.7.17
- * @release         2026-08-22
+ * @version         1.7.18
+ * @release         2026-08-25
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/