+++ webapp/utils/body-limit.js
@@ -5,8 +5,8 @@
  *                  parsers ahead of the global default, and turn oversize bodies into the
  *                  framework JSON error envelope.
  * @file            webapp/utils/body-limit.js
- * @version         1.7.17
- * @release         2026-08-22
+ * @version         1.7.18
+ * @release         2026-08-25
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025-2026 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/